requirements.txt
bson==0.5.10
certifi==2022.9.24       
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.32.0       
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
npm==0.1.1
optional-django==0.1.0
pipenv==2022.11.11
platformdirs==2.5.4
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.1
six==1.16.0
template-render==0.1.0
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
Werkzeug==2.2.2